requirements.txt
ipykernel
azure-cognitiveservices-speech
azure-core
azure-common
azure-ai-textanalytics